retrieveOpenAIContainer function move to lib
